+++ examples/Threads/future2.cpp
@@ -37,7 +37,7 @@
 
 #if defined (ACE_HAS_THREADS)
 
-typedef ACE_Atomic_Op<ACE_Thread_Mutex, u_long> ATOMIC_INT;
+typedef ACE_Atomic_Op<ACE_Thread_Mutex, int> ATOMIC_INT;
 
 // a counter for the tasks..
 static ATOMIC_INT scheduler_open_count (0);
@@ -507,7 +507,7 @@ main (int, char *[])
 }
 
 #if defined (ACE_TEMPLATES_REQUIRE_SPECIALIZATION)
-template class ACE_Atomic_Op<ACE_Thread_Mutex, u_long>;
+template class ACE_Atomic_Op<ACE_Thread_Mutex, int>;
 #endif /* ACE_TEMPLATES_REQUIRE_SPECIALIZATION */
 
 #else
